+++ core/src/test/java/io/cucumber/core/plugin/NoPublishFormatterTest.java
@@ -17,19 +17,22 @@ class NoPublishFormatterTest {
         noPublishFormatter.setMonochrome(true);
         noPublishFormatter.printBanner();
         assertThat(bytes, isBytesEqualTo("" +
-                "┌─────────────────────────────────────────────────────────────────────────────┐\n" +
-                "│ Share your Cucumber Report with your team at https://reports.cucumber.io    │\n" +
-                "│ Activate publishing with one of the following:                              │\n" +
-                "│                                                                             │\n" +
-                "│ src/test/resources/cucumber.properties:    cucumber.publish.enabled=true    │\n" +
-                "│ Environment variable:                      CUCUMBER_PUBLISH_ENABLED=true    │\n" +
-                "│ JUnit:                                     @CucumberOptions(publish = true) │\n" +
-                "│                                                                             │\n" +
-                "│ More information at https://reports.cucumber.io/docs/cucumber-jvm           │\n" +
-                "│                                                                             │\n" +
-                "│ To disable this message, add cucumber.publish.quiet=true to                 │\n" +
-                "│ src/test/resources/cucumber.properties                                      │\n" +
-                "└─────────────────────────────────────────────────────────────────────────────┘\n"));
+                "┌───────────────────────────────────────────────────────────────────────────────────┐\n" +
+                "│ Share your Cucumber Report with your team at https://reports.cucumber.io          │\n" +
+                "│ Activate publishing with one of the following:                                    │\n" +
+                "│                                                                                   │\n" +
+                "│ src/test/resources/cucumber.properties:          cucumber.publish.enabled=true    │\n" +
+                "│ src/test/resources/junit-platform.properties:    cucumber.publish.enabled=true    │\n" +
+                "│ Environment variable:                            CUCUMBER_PUBLISH_ENABLED=true    │\n" +
+                "│ JUnit:                                           @CucumberOptions(publish = true) │\n" +
+                "│                                                                                   │\n" +
+                "│ More information at https://reports.cucumber.io/docs/cucumber-jvm                 │\n" +
+                "│                                                                                   │\n" +
+                "│ Disable this message with one of the following:                                   │\n" +
+                "│                                                                                   │\n" +
+                "│ src/test/resources/cucumber.properties:          cucumber.publish.quiet=true      │\n" +
+                "│ src/test/resources/junit-platform.properties:    cucumber.publish.quiet=true      │\n" +
+                "└───────────────────────────────────────────────────────────────────────────────────┘\n"));
     }
 
 }