+++ core/src/main/java/io/cucumber/core/plugin/UrlReporter.java
@@ -3,9 +3,6 @@ package io.cucumber.core.plugin;
 import io.cucumber.plugin.ColorAware;
 
 import java.io.PrintStream;
-import java.net.URL;
-
-import static java.util.Arrays.asList;
 
 final class UrlReporter implements ColorAware {
     private final PrintStream out;
@@ -15,16 +12,11 @@ final class UrlReporter implements ColorAware {
         this.out = out;
     }
 
-    public void report(URL url) {
-        String reportUrl = String.format("https://reports.cucumber.io%s", url.getPath());
-
-        Banner banner = new Banner(out, monochrome);
-        banner.print(asList(
-            new Banner.Line("View your Cucumber Report at:"),
-            new Banner.Line(reportUrl, AnsiEscapes.CYAN, AnsiEscapes.INTENSITY_BOLD, AnsiEscapes.UNDERLINE),
-            new Banner.Line(""),
-            new Banner.Line("This report will self-destruct in 24h unless it is claimed or deleted.")),
-            AnsiEscapes.GREEN, AnsiEscapes.INTENSITY_BOLD);
+    public void report(String message) {
+        if (monochrome) {
+            message = message.replaceAll("\u001B\\[[;\\d]*m", "");
+        }
+        out.print(message);
     }
 
     @Override